Component-ised the main site
- Changed most things to be in a component.
- I will wait to see Lucas's new code for the logo cloud before I make that into a component
