+++ mlai-frontend/src/components/footer.tsx
@@ -1,0 +1,156 @@
+
+const footerNavigation = {
+    events: [
+      { name: 'Meetups', href: '/#events' },
+      { name: 'Hackathons', href: '/#events' },
+      { name: 'Codecamp', href: '#' },
+    ],
+    volunteering: [
+      { name: 'How to volunteer', href: '/contact' },
+      { name: 'Apply', href: '/contact' },
+    ],
+    Sponsoring: [
+      { name: 'Info for sponsors', href: '/sponsors' },
+    ],
+    About: [
+      { name: 'Contact', href: '/#about' },
+      { name: 'Code of Conduct', href: '/codeofconduct' },
+      { name: 'LinkedIn', href: 'https://www.linkedin.com/company/mlai-aus-inc' },
+    ],
+    social: [
+      {
+        name: 'LinkedIn',
+        href: 'https://www.linkedin.com/company/mlai-aus-inc',
+        icon: (props: any) => (
+          <svg
+            xmlns="http://www.w3.org/2000/svg"
+            className="h-5 w-5"
+            fill="currentColor"
+            viewBox="0 0 24 24">
+            <path
+              d="M4.98 3.5c0 1.381-1.11 2.5-2.48 2.5s-2.48-1.119-2.48-2.5c0-1.38 1.11-2.5 2.48-2.5s2.48 1.12 2.48 2.5zm.02 4.5h-5v16h5v-16zm7.982 0h-4.968v16h4.969v-8.399c0-4.67 6.029-5.052 6.029 0v8.399h4.988v-10.131c0-7.88-8.922-7.593-11.018-3.714v-2.155z" />
+          </svg>
+        ),
+      },
+      {
+        name: 'Instagram',
+        href: 'https://www.linkedin.com/company/mlai-aus-inc',
+        icon: (props: any) => (
+          <svg fill="currentColor" viewBox="0 0 24 24" {...props}>
+            <path
+              fillRule="evenodd"
+              d="M12.315 2c2.43 0 2.784.013 3.808.06 1.064.049 1.791.218 2.427.465a4.902 4.902 0 011.772 1.153 4.902 4.902 0 011.153 1.772c.247.636.416 1.363.465 2.427.048 1.067.06 1.407.06 4.123v.08c0 2.643-.012 2.987-.06 4.043-.049 1.064-.218 1.791-.465 2.427a4.902 4.902 0 01-1.153 1.772 4.902 4.902 0 01-1.772 1.153c-.636.247-1.363.416-2.427.465-1.067.048-1.407.06-4.123.06h-.08c-2.643 0-2.987-.012-4.043-.06-1.064-.049-1.791-.218-2.427-.465a4.902 4.902 0 01-1.772-1.153 4.902 4.902 0 01-1.153-1.772c-.247-.636-.416-1.363-.465-2.427-.047-1.024-.06-1.379-.06-3.808v-.63c0-2.43.013-2.784.06-3.808.049-1.064.218-1.791.465-2.427a4.902 4.902 0 011.153-1.772A4.902 4.902 0 015.45 2.525c.636-.247 1.363-.416 2.427-.465C8.901 2.013 9.256 2 11.685 2h.63zm-.081 1.802h-.468c-2.456 0-2.784.011-3.807.058-.975.045-1.504.207-1.857.344-.467.182-.8.398-1.15.748-.35.35-.566.683-.748 1.15-.137.353-.3.882-.344 1.857-.047 1.023-.058 1.351-.058 3.807v.468c0 2.456.011 2.784.058 3.807.045.975.207 1.504.344 1.857.182.466.399.8.748 1.15.35.35.683.566 1.15.748.353.137.882.3 1.857.344 1.054.048 1.37.058 4.041.058h.08c2.597 0 2.917-.01 3.96-.058.976-.045 1.505-.207 1.858-.344.466-.182.8-.398 1.15-.748.35-.35.566-.683.748-1.15.137-.353.3-.882.344-1.857.048-1.055.058-1.37.058-4.041v-.08c0-2.597-.01-2.917-.058-3.96-.045-.976-.207-1.505-.344-1.858a3.097 3.097 0 00-.748-1.15 3.098 3.098 0 00-1.15-.748c-.353-.137-.882-.3-1.857-.344-1.023-.047-1.351-.058-3.807-.058zM12 6.865a5.135 5.135 0 110 10.27 5.135 5.135 0 010-10.27zm0 1.802a3.333 3.333 0 100 6.666 3.333 3.333 0 000-6.666zm5.338-3.205a1.2 1.2 0 110 2.4 1.2 1.2 0 010-2.4z"
+              clipRule="evenodd"
+            />
+          </svg>
+        ),
+      },
+      {
+        name: 'Slack',
+        href: 'https://join.slack.com/t/aihackmelb/shared_invite/zt-2br01q4n9-PiiT9mgEOPMEG__8SOWQ8g',
+        icon: (props: any) => (
+          <svg
+            className="h-6 w-6"
+            fill="currentColor"
+            viewBox="0 0 24 24"
+            xmlns="http://www.w3.org/2000/svg"
+            fillRule="evenodd"
+            clipRule="evenodd">
+            <path
+              d="M22.672 15.226l-2.432.811.841 2.515c.33 1.019-.209 2.127-1.23 2.456-1.15.325-2.148-.321-2.463-1.226l-.84-2.518-5.013 1.677.84 2.517c.391 1.203-.434 2.542-1.831 2.542-.88 0-1.601-.564-1.86-1.314l-.842-2.516-2.431.809c-1.135.328-2.145-.317-2.463-1.229-.329-1.018.211-2.127 1.231-2.456l2.432-.809-1.621-4.823-2.432.808c-1.355.384-2.558-.59-2.558-1.839 0-.817.509-1.582 1.327-1.846l2.433-.809-.842-2.515c-.33-1.02.211-2.129 1.232-2.458 1.02-.329 2.13.209 2.461 1.229l.842 2.515 5.011-1.677-.839-2.517c-.403-1.238.484-2.553 1.843-2.553.819 0 1.585.509 1.85 1.326l.841 2.517 2.431-.81c1.02-.33 2.131.211 2.461 1.229.332 1.018-.21 2.126-1.23 2.456l-2.433.809 1.622 4.823 2.433-.809c1.242-.401 2.557.484 2.557 1.838 0 .819-.51 1.583-1.328 1.847m-8.992-6.428l-5.01 1.675 1.619 4.828 5.011-1.674-1.62-4.829z" />
+          </svg>
+        ),
+      },
+    ],
+  }
+
+export default function Events() {
+    return (
+        <footer className="bg-black" aria-labelledby="footer-heading">
+            <h2 id="footer-heading" className="sr-only">
+                Footer
+            </h2>
+            <div className="mx-auto max-w-7xl px-6 pb-8 pt-6 sm:pt-24 lg:px-8 lg:pt-8">
+                <div className="xl:grid xl:grid-cols-3 xl:gap-8">
+                    <div className="grid grid-cols-2 gap-8 xl:col-span-2">
+                        <div className="md:grid md:grid-cols-2 md:gap-8">
+                            <div>
+                                <h3 className="text-sm font-semibold leading-6 text-white">Events</h3>
+                                <ul role="list" className="mt-6 space-y-4">
+                                    {footerNavigation.events.map((item) => (
+                                        <li key={item.name}>
+                                            <a href={item.href} className="text-sm leading-6 text-gray-300 hover:text-white">
+                                                {item.name}
+                                            </a>
+                                        </li>
+                                    ))}
+                                </ul>
+                            </div>
+                            <div className="mt-10 md:mt-0">
+                                <h3 className="text-sm font-semibold leading-6 text-white">Volunteering</h3>
+                                <ul role="list" className="mt-6 space-y-4">
+                                    {footerNavigation.volunteering.map((item) => (
+                                        <li key={item.name}>
+                                            <a href={item.href} className="text-sm leading-6 text-gray-300 hover:text-white">
+                                                {item.name}
+                                            </a>
+                                        </li>
+                                    ))}
+                                </ul>
+                            </div>
+                        </div>
+                        <div className="md:grid md:grid-cols-2 md:gap-8">
+                            <div>
+                                <h3 className="text-sm font-semibold leading-6 text-white">Sponsoring</h3>
+                                <ul role="list" className="mt-6 space-y-4">
+                                    {footerNavigation.Sponsoring.map((item) => (
+                                        <li key={item.name}>
+                                            <a href={item.href} className="text-sm leading-6 text-gray-300 hover:text-white">
+                                                {item.name}
+                                            </a>
+                                        </li>
+                                    ))}
+                                </ul>
+                            </div>
+                            <div className="mt-10 md:mt-0">
+                                <h3 className="text-sm font-semibold leading-6 text-white">About</h3>
+                                <ul role="list" className="mt-6 space-y-4">
+                                    {footerNavigation.About.map((item) => (
+                                        <li key={item.name}>
+                                            <a href={item.href} className="text-sm leading-6 text-gray-300 hover:text-white">
+                                                {item.name}
+                                            </a>
+                                        </li>
+                                    ))}
+                                </ul>
+                            </div>
+                        </div>
+                    </div>
+
+                    <div className="flex justify-center items-center p-4 pt-8">
+                        <img
+                            className="h-10 w-auto"
+                            src="/text_logo.png"
+                            alt="MLAI text logo"
+                        />
+                    </div>
+
+
+                </div>
+                <div className="mt-8 border-t border-gray-400 pt-8 sm:mt-10 md:flex md:items-center md:justify-between lg:mt-12">
+                    <div className="flex space-x-6 md:order-2">
+                        {footerNavigation.social.map((item) => (
+                            <a key={item.name} href={item.href} className="text-gray-300 hover:text-gray-400">
+                                <span className="sr-only">{item.name}</span>
+                                <item.icon className="h-6 w-6" aria-hidden="true" />
+                            </a>
+                        ))}
+                    </div>
+                    <p className="mt-8 text-xs leading-5 text-gray-300 md:order-1 md:mt-0">
+                        &copy; 2024 MLAI Aus Inc. All rights reserved.
+                    </p>
+                </div>
+            </div>
+        </footer>
+    )
+}
